+++ include/settings.hpp
@@ -10,6 +10,7 @@
 #include "scheduling/circular_drr_scheduler.hpp"
 #include "scheduling/default_drr_scheduler.hpp"
 #include "standard_info.hpp"
+#include "types.hpp"
 
 class Settings {
     public:
@@ -23,7 +24,7 @@ class Settings {
             int packet_count, 
             int packet_size, 
             int queue_count, 
-            int queue_quant,
+            double queue_quant,
             int queue_limit,
             double time_lambda);
 
@@ -37,7 +38,7 @@ class Settings {
         int get_packet_count();
         int get_packet_size();
         int get_queue_count();
-        int get_queue_quant();
+        double get_queue_quant();
         int get_queue_limit();
         double get_time_lambda();
         int get_resource_block_per_tti_limit();
@@ -51,7 +52,7 @@ class Settings {
         int packet_count; // Количество пакетов для каждого повтора
         int packet_size; // Размеры пакетов для каждого повтора
         int queue_count; // Количество очередей для каждого повтора
-        int queue_quant; // Квант времени очередей для каждого повтора
+        double queue_quant; // Квант времени очередей для каждого повтора
         int queue_limit; // Размеры очередей для каждого повтора
         double time_lambda; // Частота прихода пакетов в единицу времени
 };
